module2/m2hw15FINAL/requirements.txt
aiofile==3.5.1
aiohttp==3.7.4.post0
aiopath==0.5.8
alembic==1.6.5
anyio==3.2.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
Automat==20.2.0
caio==0.8.4
cffi==1.14.6
chardet==4.0.0
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.8
cssselect==1.1.0
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
inflection==0.5.1
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
multidict==5.1.0
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.1
queuelib==1.6.2
Scrapy==2.5.0
scrapy-user-agents==0.1.1
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.21
sqlparse==0.4.1
Twisted==21.7.0
twisted-iocpsupport==1.0.1
typing-extensions==3.10.0.2
ua-parser==0.10.0
user-agents==2.2.0
w3lib==1.22.0
Werkzeug==2.0.1
WTForms==2.3.3
yarl==1.6.3
zope.interface==5.4.0
